+++ CHANGES.md
@@ -25,6 +25,7 @@ Changes from ns-3.38 to ns-3-dev
 * (lr-wpan) Added `LrwpanMac::MlmeGetRequest` function and the corresponding confirm callbacks as well as `LrwpanMac::SetMlmeGetConfirm` function.
 * (applications) Added `Tx` and `TxWithAddresses` trace sources in `UdpClient`.
 * (stats) Added `Histogram::Clear` function to clear the histogram contents.
+* (flow-monitor) Added `FlowMonitor::ResetAllStats` function to reset the FlowMonitor statistics.
 
 ### Changes to existing API